AI News Summary Tool
This tool is for AutoProphet News AI Summarization. In order for it to be used an api key secret must be present under the name "Gemini_Key". Articles need to be delivered to the "USER_Gemini_Prompt" variable. This is usable as a back-end product and only requires an interface.