Uploading Final Report with Demo Link attached
Uploading a PDF version of the final report for better readability and in the Appendix section L it has the link to the demo video explaining how to use the backend portion of the project. The file was too large to upload individually on GitHub. This will be the last and final update that I will do for this project. Good luck upcoming semester class!